CheckerService/requirements.txt
flask
telethon
httpx
requests
python-dotenv
aiogram
uvicorn
quart
hypercorn